+++ src/webapp/src/store/modules/orchestration.js
@@ -122,12 +122,6 @@ const getters = {
 }
 
 const actions = {
-  addConfigurationProfile(_, profile) {
-    return orchestrationsApi.addConfigurationProfile(profile).catch(error => {
-      Vue.toasted.global.error(error)
-    })
-  },
-
   createSubscription(_, subscription) {
     return orchestrationsApi.createSubscription(subscription)
   },
@@ -324,15 +318,13 @@ const mutations = {
     )
     configuration.settings.forEach(setting => {
       const isIso8601Date = setting.kind && setting.kind === 'date_iso8601'
-      configuration.profiles.forEach(profile => {
-        const isDefaultNeeded =
-          profile.config.hasOwnProperty(setting.name) &&
-          profile.config[setting.name] === null &&
-          requiredSettingsKeys.includes(setting.name)
-        if (isIso8601Date && isDefaultNeeded) {
-          profile.config[setting.name] = utils.getFirstOfMonthAsYYYYMMDD()
-        }
-      })
+      const isDefaultNeeded =
+        configuration.config.hasOwnProperty(setting.name) &&
+        configuration.config[setting.name] === null &&
+        requiredSettingsKeys.includes(setting.name)
+      if (isIso8601Date && isDefaultNeeded) {
+        configuration.config[setting.name] = utils.getFirstOfMonthAsYYYYMMDD()
+      }
     })
     state[target] = configuration
   },