+++ tests/meltano/core/test_elt_context.py
@@ -8,7 +8,6 @@ from meltano.core.plugin import PluginType, Variant
 def assert_extractor_env(extractor, env):
     assert env["MELTANO_EXTRACTOR_NAME"] == extractor.name
     assert env["MELTANO_EXTRACTOR_NAMESPACE"] == extractor.definition.namespace
-    assert env["MELTANO_EXTRACTOR_PROFILE"] == extractor.plugin.current_profile_name
     assert env["MELTANO_EXTRACTOR_VARIANT"] == extractor.definition.current_variant_name
 
     assert env["MELTANO_EXTRACT_TEST"] == env["TAP_MOCK_TEST"] == "mock"
@@ -18,7 +17,6 @@ def assert_extractor_env(extractor, env):
 def assert_loader_env(loader, env):
     assert env["MELTANO_LOADER_NAME"] == loader.name
     assert env["MELTANO_LOADER_NAMESPACE"] == loader.definition.namespace
-    assert env["MELTANO_LOADER_PROFILE"] == loader.plugin.current_profile_name
     assert env["MELTANO_LOADER_VARIANT"] == loader.definition.current_variant_name
 
     assert (
@@ -37,7 +35,6 @@ def assert_loader_env(loader, env):
 def assert_transform_env(transform, env):
     assert env["MELTANO_TRANSFORM_NAME"] == transform.name
     assert env["MELTANO_TRANSFORM_NAMESPACE"] == transform.definition.namespace
-    assert env["MELTANO_TRANSFORM_PROFILE"] == transform.plugin.current_profile_name
     assert env["MELTANO_TRANSFORM_VARIANT"] == Variant.ORIGINAL_NAME
 
     assert env["MELTANO_TRANSFORM__PACKAGE_NAME"] == "dbt_mock"
@@ -46,7 +43,6 @@ def assert_transform_env(transform, env):
 def assert_transformer_env(transformer, env):
     assert env["MELTANO_TRANSFORMER_NAME"] == transformer.name
     assert env["MELTANO_TRANSFORMER_NAMESPACE"] == transformer.definition.namespace
-    assert env["MELTANO_TRANSFORMER_PROFILE"] == transformer.plugin.current_profile_name
     assert env["MELTANO_TRANSFORMER_VARIANT"] == Variant.ORIGINAL_NAME
 
     assert (